web-animations/animation-player/animation-player-timeline.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationPlayer timeline attribute tests</title>
<meta name="assert" content="The timeline associated with this player. Setting this attribute updates the object’s timeline using the procedure to set the timeline of a player">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationplayer-timeline">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var player = document.timeline.play(newAnimation(createDiv(this)));

    assert_not_equals(player.timeline, null, 'AnimationPlayer.timeline attribute should ' +
        'not be null');
    assert_equals(player.timeline, document.timeline, 'AnimationPlayer.timeline ' +
        'should be equal to document.timeline');
}, 'The AnimationPlayer.timeline attribute returns the timeline, used to create the player ' +
    'by AnimationTimeline.play()');

test(function() {
    var player = new AnimationPlayer(null, document.timeline);

    assert_not_equals(player.timeline, null, 'AnimationPlayer.timeline attribute should ' +
        'not be null');
    assert_equals(player.timeline, document.timeline, 'AnimationPlayer.timeline ' +
        'should be equal to document.timeline');
}, 'The AnimationPlayer.timeline attribute returns the timeline, passed to ' +
    'AnimationPlayer constructor ');

async_test(function(t) {
    var player = document.timeline.play(newAnimation(createDiv(t)));
    var readyPromise = player.ready;

    player.ready.then(
        //onFulfilled. This never should called because we are going to reject
        // current ready promise and cancel pending play task
        t.unreached_func('Current ready promise should be rejected'),
        // onRejected. Should be called with AbortError
        t.step_func(function(error) {
            assert_class_string(error, 'AbortError', 'Promise should be rejected ' +
                'with a DOMException named \'AbortError\'');
            t.done();
        })
    );

    player.timeline = null;
    assert_not_equals(player.ready, readyPromise, 'Player\'s current ready promise ' +
        'should be a new resolved Promise object');
}, 'Test that setting value of AnimationPlayer.timeline attribute to null ' +
    'resets a player\'s pending tasks');

async_test(function(t) {
    var player = document.timeline.play(newAnimation(createDiv(t)));
    player.pause();
    var doc = newHTMLDocument();
    player.timeline = doc.timeline;

    player.ready.then(t.step_func(function() {
        assert_equals(player.timeline, doc.timeline, 'Player should have new timeline');
        assert_equals(player.playState, 'paused', 'Paused task should not be canceled');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that setting value of AnimationPlayer.timeline attribute to new not null value ' +
    'don\'t cancel pending play and pause tasks');

async_test(function(t) {
    var player = document.timeline.play(newAnimation(createDiv(t)));
    player.finish();
    var doc = newHTMLDocument();
    player.timeline = doc.timeline;

    player.ready.then(t.step_func(function() {
        assert_equals(player.timeline, doc.timeline, 'Player should have new timeline');
        assert_equals(player.playState, 'finished', 'Player finished state ' +
            'should be preserved');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that setting value of AnimationPlayer.timeline attribute to new not null value ' +
    'does not change player finished state');

async_test(function(t) {
    var numOfCallsNullTime = 0;
    var source = new Animation(createDiv(t),
        function(timeFraction, currentTarget, animationObject) {
            if (timeFraction === null) {
                numOfCallsNullTime++;
            }
        }, 100);
    var player = document.timeline.play(source);
    var doc = newHTMLDocument();
    player.timeline = doc.timeline;

    player.ready.then(t.step_func(function() {
        setTimeout(t.step_func(function() {
            assert_equals(numOfCallsNullTime, 1, 'AnimationEffect callback should ' +
                'be called with null timeFraction after changing of player source');
            t.done();
        }), 50);
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that EffectCallback is called with null timeFraction ' +
    'if player timeline is changed to new not null value');
</script>
</body>
